Refactor: Paging to query between dates
Querydsl은 group by 에서 paging을 지원하지 않음..
두 날짜 사이 주문기록을 가져오도록 리팩토링

footer: OrderDetailService, controller 미완성

+++ src/main/java/shop/S5G/shop/dto/order/OrderWithDetailResponseDto.java
@@ -11,6 +11,6 @@ public record OrderWithDetailResponseDto(
     long totalPrice,
     // 리스트에서 노출시킬 대표 디테일("'예시타이틀' 2권 외...")
     String representTitle,
-    int totalKind,      // 종류(총 N종)
+    long totalKind,      // 종류(총 N종)
     int totalQuantity   // 양(총 N권)
 ) {}
